+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Translations: Added Bulgarian translation thanks to Nikolay Stoychev.
 - Timeregistration: Activity name and code visible on ediding time registration
 - Core: Template and function custom field use text values instead of id's of select fields
+- Core: Turn off GarbageCollection while upgrading
 
 13-10-2020 6.4.181
 - Core: In system settings, extra filter 'Disabled users', make sortable by change date